+++ lib/pages/error.ex
@@ -1,0 +1,4 @@
+defmodule Pages.Error do
+  @moduledoc "Errors that are specifically raised from page drivers."
+  defexception [:message]
+end